backend/ai/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
auraloss==0.3.0
boto3==1.26.95
botocore==1.29.95
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.2.13
django-cors-headers
django-rest-swagger==2.2.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-yasg
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.10.0
flatbuffers==23.3.3
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.12.0
h2==4.1.0
h5py==3.8.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.4
itypes==1.2.0
jax==0.4.6
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.12.0
libclang==15.0.6.1
librosa==0.8.1
llvmlite==0.38.1
Markdown==3.4.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
norbert==0.2.1
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.2
openapi-codec==1.3.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.2.0
pooch==1.7.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.0
pydub==0.25.1
pyloudnorm==0.1.0
python-dateutil==2.8.2
pytorch-lightning==1.3.2
pytz==2022.7.1
PyYAML==5.4.1
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
simplejson==3.18.4
six==1.16.0
sniffio==1.3.0
SoundFile==0.10.3.post1
sox==1.4.1
spleeter==2.3.2
sqlparse==0.4.3
sympy==1.11.1
termcolor==2.2.0
threadpoolctl==3.1.0
torch==2.0.0
torchaudio==2.0.1
torchmetrics==0.6.0
torchvision==0.15.1
tqdm==4.65.0
typer==0.3.2
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
